requirements.txt
transformers
torch
biopython
fastapi
uvicorn
streamlit
